Rename AI import to tool-calling in actions directory
