+++ app/layout.tsx
@@ -3,8 +3,8 @@ import { Geist, Geist_Mono } from "next/font/google";
 import { Analytics } from "@vercel/analytics/next";
 import { siteConfig } from "@/lib/site-config";
 import { Toast } from "@/components/ui/toast";
+import { AI } from "@/actions/tool-calling";
 import { Providers } from "@/lib/providers";
-import { AI } from "@/app/actions";
 import { cn } from "@/lib/utils";
 import { Suspense } from "react";
 import "./globals.css";
